requirements.txt
openai
python-dotenv
ffmpeg-python